Refactor code comments in data loader builder and MNIST dataset
This commit refactors the code comments in the `data_loader_builder.rs` file and the `mnist.rs` file. In the `data_loader_builder.rs` file, the comment for the `new` function has been updated to provide clearer information about the arguments and return value. In the `mnist.rs` file, a typo has been fixed in a comment that describes the location of the dataset files.